Corrected channel selection condition in energy calculation
This commit addresses a logical error in the channel selection condition when calculating energy. 

Previously, the channel selection for `board` other than 3 was not properly handled, causing the condition to be valid even if the `board` was 3. This issue lead to incorrect energy calculations if the `channel` happened to be the same as `backMostRingForOthers`, even though the `board` was 3.

The code was modified to ensure that the `channel` equals `backMostRingForOthers` only when the `board` is not 3. This change should ensure accurate energy calculations based on the respective `channel` and `board` conditions.

Furthermore, the energyThreshold has been raised to 1500 and the upper limit of the histogram has been raised to 16384, which is the maximum the CAEN digitizer can have as of the time of this commit.